refactor dev options
